+++ build.sh
@@ -3,7 +3,7 @@
 ##################################################################################
 # Custom build tool for Realm Objective-C binding.
 #
-# (C) Copyright 2011-2015 by realm.io.
+# (C) Copyright 2011-2022 by realm.io.
 ##################################################################################
 
 # Warning: pipefail is not a POSIX compatible option, but on macOS it works just fine.
@@ -87,7 +87,9 @@ environment variables:
   CONFIGURATION: Debug or Release (default)
   REALM_CORE_VERSION: version in x.y.z format or "current" to use local build
   REALM_EXTRA_BUILD_ARGUMENTS: additional arguments to pass to the build tool
-  REALM_XCODE_VERSION: the version number of Xcode to use (e.g.: 8.1)
+  REALM_XCODE_VERSION: the version number of Xcode to use (e.g.: 13.3.1)
+  REALM_XCODE_OLDEST_VERSION: the version number of oldest available Xcode to use (e.g.: 12.4)
+  REALM_XCODE_LATEST_VERSION: the version number of latest available Xcode to use (e.g.: 13.3.1)
 EOF
 }
 
@@ -439,6 +441,39 @@ case "$COMMAND" in
         find build/DerivedData/Realm/Build/Products -name 'RealmSwift.framework' \
             | sed 's/.*/-framework &/' \
             | xargs xcodebuild -create-xcframework -allow-internal-distribution -output build/RealmSwift.xcframework
+
+        # Because we have a module named Realm and a type named Realm we need to manually resolve the naming
+        # collisions that are happening. These collisions create a red herring which tells the user the xcframework
+        # was compiled with an older Swift version and is not compatible with the current compiler.
+        find build/RealmSwift.xcframework -name "*.swiftinterface" -exec sed -i '' 's/Realm\.//g' {} \; \
+        -exec sed -i '' 's/import Private/import Realm.Private/g' {} \; \
+        -exec sed -i '' 's/RealmSwift.Configuration/RealmSwift.Realm.Configuration/g' {} \; \
+        -exec sed -i '' 's/extension Configuration/extension Realm.Configuration/g' {} \; \
+        -exec sed -i '' 's/RealmSwift.Error/RealmSwift.Realm.Error/g' {} \; \
+        -exec sed -i '' 's/extension Error/extension Realm.Error/g' {} \; \
+        -exec sed -i '' 's/RealmSwift.AsyncOpenTask/RealmSwift.Realm.AsyncOpenTask/g' {} \; \
+        -exec sed -i '' 's/RealmSwift.UpdatePolicy/RealmSwift.Realm.UpdatePolicy/g' {} \; \
+        -exec sed -i '' 's/RealmSwift.Notification /RealmSwift.Realm.Notification /g' {} \; \
+        -exec sed -i '' 's/RealmSwift.Notification,/RealmSwift.Realm.Notification,/g' {} \; \
+        -exec sed -i '' 's/τ_1_0/V/g' {} \; # Generics will use τ_1_0 which needs to be changed to the correct type name.
+
+        exit 0
+        ;;
+
+    "verify-xcframework-evolution-mode")
+        export REALM_EXTRA_BUILD_ARGUMENTS="$REALM_EXTRA_BUILD_ARGUMENTS REALM_BUILD_LIBRARY_FOR_DISTRIBUTION=YES"
+        # set the Xcode version to the oldest
+        export REALM_XCODE_VERSION=$REALM_XCODE_OLDEST_VERSION
+        unset REALM_SWIFT_VERSION
+        sh build.sh xcframework ios
+        # copy the xcframework to the testing target
+        rm -rf examples/installation/xcframework-evolution
+        mkdir examples/installation/xcframework-evolution
+        cp -r build/*.xcframework examples/installation/xcframework-evolution
+        export REALM_XCODE_VERSION=$REALM_XCODE_LATEST_VERSION
+        unset REALM_SWIFT_VERSION
+        cd examples/installation
+        sh build.sh "test-ios-swift-xcframework"
 
         exit 0
         ;;